feat: RestaurantReviewStatService를 제거하고 RestaurantService로 통일
